Remove cycles due to public/module/k8s/index.ts
