+++ frontend/packages/ceph-storage-plugin/src/plugin.ts
@@ -27,7 +27,7 @@ import {
 import { ClusterServiceVersionModel } from '@console/operator-lifecycle-manager/src/models';
 import { GridPosition } from '@console/shared/src/components/dashboard/DashboardGrid';
 import { OverviewQuery } from '@console/internal/components/dashboard/dashboards-page/cluster-dashboard/queries';
-import { referenceForModel, referenceFor } from '@console/internal/module/k8s';
+import { referenceForModel, referenceFor } from '@console/internal/module/k8s/k8s';
 import { PersistentVolumeClaimModel } from '@console/internal/models';
 import { getCephHealthState } from './components/dashboard-page/storage-dashboard/status-card/utils';
 import { isClusterExpandActivity } from './components/dashboard-page/storage-dashboard/activity-card/cluster-expand-activity';
