+++ frontend/packages/dev-console/src/plugin.tsx
@@ -21,7 +21,7 @@ import {
 import { NamespaceRedirect } from '@console/internal/components/utils/namespace-redirect';
 import { CodeIcon } from '@patternfly/react-icons';
 import { FLAGS } from '@console/shared/src/constants';
-import { referenceForModel } from '@console/internal/module/k8s';
+import { referenceForModel } from '@console/internal/module/k8s/k8s';
 import * as models from './models';
 import { getKebabActionsForKind } from './utils/kebab-actions';
 import {
