+++ frontend/packages/network-attachment-definition-plugin/src/plugin.ts
@@ -9,7 +9,7 @@ import {
   ModelDefinition,
   RoutePage,
 } from '@console/plugin-sdk';
-import { referenceForModel } from '@console/internal/module/k8s';
+import { referenceForModel } from '@console/internal/module/k8s/k8s';
 import { FLAG_KUBEVIRT } from '@console/kubevirt-plugin/src/plugin';
 import * as models from './models';
 import { NetworkAttachmentDefinitionsYAMLTemplates } from './models/templates';
