+++ frontend/packages/noobaa-storage-plugin/src/plugin.ts
@@ -13,7 +13,7 @@ import {
   YAMLTemplate,
 } from '@console/plugin-sdk';
 import { GridPosition } from '@console/shared/src/components/dashboard/DashboardGrid';
-import { referenceForModel } from '@console/internal/module/k8s';
+import { referenceForModel } from '@console/internal/module/k8s/k8s';
 import { ClusterServiceVersionModel } from '@console/operator-lifecycle-manager';
 import * as models from './models';
