--- frontend/public/module/k8s/event.ts
@@ -1,7 +1,0 @@
-export type EventInvolvedObject = {
-  apiVersion?: string;
-  kind?: string;
-  name?: string;
-  uid?: string;
-  namespace?: string;
-};
