+++ frontend/public/reducers/features.ts
@@ -15,7 +15,7 @@ import {
   MachineModel,
   PrometheusModel,
 } from '../models';
-import { referenceForModel } from '../module/k8s';
+import { referenceForModel } from '../module/k8s/k8s';
 import { RootState } from '../redux';
 import { ActionType as K8sActionType } from '../actions/k8s';
 import { FeatureAction, ActionType } from '../actions/features';
